+++ tests/ElasticaProviderBundle/Command/ProvideCommandTest.php
@@ -29,23 +29,15 @@ class ProvideCommandTest extends \PHPUnit_Framework_TestCase
         $command = $application->find('elasticsearch:provide');
         $this->commandTester = new CommandTester($command);
 
-        $this->handler = $this
-            ->getMockBuilder(Handler::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
+        $this->handler = $this->prophesize(Handler::class);
 
         $container = new Container();
-        $container->set('gbprod.elastica_provider.handler', $this->handler);
+        $container->set('gbprod.elastica_provider.handler', $this->handler->reveal());
         $container->set('event_dispatcher', $this->getMock(EventDispatcherInterface::class));
 
-        $this->client = $this
-            ->getMockBuilder(Client::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
+        $this->client = $this->prophesize(Client::class);
 
-        $container->set('gbprod.elastica_provider.default_client', $this->client);
+        $container->set('gbprod.elastica_provider.default_client', $this->client->reveal());
 
         $command->setContainer($container);
     }
@@ -53,9 +45,8 @@ class ProvideCommandTest extends \PHPUnit_Framework_TestCase
     public function testExecute()
     {
         $this->handler
-            ->expects($this->once())
-            ->method('handle')
-            ->with($this->client, 'my_index', 'my_type')
+            ->handle($this->client->reveal(), 'my_index', 'my_type', null)
+            ->shouldBeCalled()
         ;
 
         $this->commandTester->execute([
@@ -67,7 +58,7 @@ class ProvideCommandTest extends \PHPUnit_Framework_TestCase
 
     public function testExecuteThrowExceptionIfClientNotFound()
     {
-        $this->setExpectedException('\InvalidArgumentException');
+        $this->setExpectedException(\InvalidArgumentException::class);
 
         $this->commandTester->execute([
             'command'  => 'elasticsearch:provide',
@@ -77,4 +68,31 @@ class ProvideCommandTest extends \PHPUnit_Framework_TestCase
         ]);
     }
 
+    public function testExecuteWithAlias()
+    {
+        $this->handler
+            ->handle($this->client->reveal(), 'my_index', 'my_type', 'my_alias')
+            ->shouldBeCalled()
+        ;
+
+        $this->commandTester->execute([
+            'command' => 'elasticsearch:provide',
+            'index'   => 'my_index',
+            'type'    => 'my_type',
+            '--alias' => 'my_alias',
+        ]);
+    }
+
+    public function testExecuteWithAliasButNoIndex()
+    {
+        $this->handler
+            ->handle($this->client->reveal(), null, null, null)
+            ->shouldBeCalled()
+        ;
+
+        $this->commandTester->execute([
+            'command' => 'elasticsearch:provide',
+            '--alias' => 'my_alias',
+        ]);
+    }
 }